I’ve refactored the tests to make use of Mock Service Worker (MSW) to mock our API endpoints. We had previously been using a bit of a hacky solution with undici which worked well enough but did not seem like the right way to do it. Had I know about MSW, I would have used that from the start.
As I’m migrating and trying to fix all the tests I’ve moved updated tests into the `spec` folder.
