+++ .travis.yml
@@ -12,7 +12,7 @@ before_script:
 
 script:
   - echo "Run full npm test when tests have been fixed"
-  - npm run lint && npm run test:unit
+  - npm run lint && npm run test
 
 after_success:
   - echo "Determine deployment environment"
