+++ package.json
@@ -15,10 +15,9 @@
         "lint:fix": "eslint --ext .ts --fix src/**",
         "lint:frontend": "eslint --ext .js src/main/assets/scripts/**",
         "lint:frontend:fix": "eslint --ext .js --fix src/main/assets/scripts/**",
-        "test": "mocha \"src/test/**/*.test.ts\" --reporter-options reportFilename=all,inlineAssets=true,reportTitle=CAS-buyer-frontend-tests-all",
+        "test": "env-cmd -f config/.env.test mocha \"src/spec/**/*.spec.ts\" --reporter-options reportFilename=all,inlineAssets=true,reportTitle=CAS-buyer-frontend-tests-all",
         "test:a11y": "mocha \"src/test/a11y/**/*.test.ts\" --timeout 60000 --reporter-options reportFilename=a11y,inlineAssets=true,reportTitle=CAS-buyer-frontend-a11y",
         "test:routes": "mocha \"src/test/routes/**/*.test.ts\" --reporter-options reportFilename=routes,inlineAssets=true,reportTitle=CAS-buyer-frontend-routes",
-        "test:unit": "env-cmd -f config/.env.test mocha \"src/test/unit/**/*.test.ts\" --reporter-options reportFilename=unit,inlineAssets=true,reportTitle=CAS-buyer-frontend-unit-tests-all",
         "test:coverage": "nyc --reporter=html npm test",
         "test:smoke": "echo 'TODO add smoke tests' && exit 1",
         "test:functional": "echo 'TODO add functional tests' && exit 1",
@@ -109,6 +108,7 @@
         "mocha-multi": "^1.1.3",
         "mocha-parallel-tests": "^2.2.2",
         "mochawesome": "^4.1.0",
+        "msw": "^1.2.2",
         "nock": "^13.3.1",
         "nyc": "^15.1.0",
         "pa11y": "^6.2.3",
@@ -118,7 +118,6 @@
         "sinon-chai": "^3.7.0",
         "supertest": "^4.0.2",
         "ts-node-dev": "^2.0.0",
-        "typescript": "^5.1.3",
-        "undici": "^5.22.1"
+        "typescript": "^5.1.3"
     }
 }
