+++ src/spec/utils/stringBuffer.spec.ts
